api/requirements.txt
accelerate==0.31.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
Django==4.2.13
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.1
fsspec==2024.6.0
h11==0.14.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12.1
tokenizers==0.19.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.6
wcwidth==0.2.13
wsproto==1.2.0
zipp==3.19.2
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
Django==4.2.13
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.1
frozenlist==1.4.1
fsspec==2024.6.0
h11==0.14.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12.1
tokenizers==0.19.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.6
wcwidth==0.2.13
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
